int to float

+++ calculator.py
@@ -1,5 +1,7 @@
 
-x = int(input("What's x? "))
-y = int(input("What's y? "))
+x = float(input("What's x? "))
+y = float(input("What's y? "))
 
 print(x + y)
+
+# print(int(input("What's x? ")) + int(input("What's y? "))) / Too complicated
